+++ include/FViz/Rendering/FVizRendering.h
@@ -3,6 +3,8 @@
 
 #include <FViz/Rendering/FVizActor.h>
 #include <FViz/Rendering/FVizCamera.h>
+#include <FViz/Rendering/FVizLookupTable.h>
+#include <FViz/Rendering/FVizMapper.h>
 #include <FViz/Rendering/FVizRenderer.h>
 #include <FViz/Rendering/FVizRenderWindow.h>
 #include <FViz/Rendering/FVizScene.h>